backend/requirements/base.txt
Django>=4.2.0,<5.0.0
django-environ>=0.11.2
djangorestframework>=3.14.0
django-cors-headers>=4.3.0
django-filter>=23.3.0
djangorestframework-simplejwt>=5.3.0
psycopg2-binary>=2.9.7
python-decouple==3.8
redis>=5.0.0
celery>=5.3.0
pandas>=2.1.0
numpy>=1.25.0
openpyxl>=3.1.0
xlrd>=2.0.0
WeasyPrint>=60.0
azure-storage-blob>=12.19.0
azure-identity>=1.15.0
msal>=1.24.1.0
gunicorn>=21.2.0
whitenoise>=6.6.0
python-decouple>=3.8
Pillow>=10.1.0
reportlab>=4.0.0
requests>=2.31.0
python-dateutil>=2.8.0
pytz>=2023.3
django-storages>=1.14.0
django-redis>=5.4.0
dj-database-url==2.1.0

# Visualization & Reports
matplotlib>=3.7.0
seaborn>=0.12.0
reportlab==4.0.7
Pillow>=10.1.0

# Web & API
requests==2.31.0
python-dateutil>=2.8.0